Handle boolean-like strings returned from redis
http://unix-jira.datapipe.net/browse/AUT-447

An issue came up where modules using the `validate_bool` function from
`stdlib` where failing compilation due to redis returning a string value
instead of a boolean for certain parameters.

This commit allows the backend to return either true or false if the
value from redis is a "true" string or a "false" string.